% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_pal.R
\name{beta_pal}
\alias{beta_pal}
\title{Beta parameters from interpretable conditions}
\usage{
beta_pal(expected_value, this_much, within)
}
\arguments{
\item{expected_value}{Expected value of beta distrbution from \link{0,1}.}

\item{this_much}{What proportion falls \code{within} the specified interval.}

\item{within}{Specify distance \code{this_much} falls within.}
}
\value{
List of parameters appropriate for \link{rbeta} family of functions.
}
\description{
Rather than knowing intrinsically what parameters are required for a
distribution, scientists tend to have a sense of what value they
\emph{expect} a measure to take, \emph{how much} they expect observations to fall
\emph{within} a certain distance of that value.
}
\details{
Visualise the distribution with \link{beta_plot}.
See \code{vignette("beta_pal")} for derivations and more information.
}
