% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plot.R
\name{beta_plot}
\alias{beta_plot}
\title{Plot beta distribution}
\usage{
beta_plot(expected_value, within, this_much, caption_width = 70)
}
\arguments{
\item{expected_value}{Expected value of beta distrbution from [0,1].}

\item{within}{Specify distance \code{this_much} falls within.}

\item{this_much}{What proportion falls \code{within} the specified interval.}

\item{caption_width}{Option to control the width of the caption, defaults to
70.}
}
\description{
This function utlises \link{beta_pal} to produce a visualisation of the beta
distribution, with user-specified interpretable conditions.
}
