% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objective.R
\name{get_objective}
\alias{get_objective}
\title{Objective function}
\usage{
get_objective(desired_mean, desired_dist, desired_mass)
}
\description{
Function that takes desrired mean, distance, and probability, and outputs
another function to be optimized.
}
\details{
This function is called by \link{beta_pal} internally.
}
